requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain and LangGraph
langchain==0.1.6
langchain-groq==0.0.1
langchain-community==0.0.20
langgraph==0.0.20
langchain-core==0.1.23

# Data Analysis
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
yfinance==0.2.36

# APIs
finnhub-python==2.4.19
google-search-results==2.4.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Groq
groq==0.4.2